fix (ngx-eq): use injection token instead of string token
