+++ projects/ngx-eq/src/ngx-eq.di-tokens.ts
@@ -1,0 +1,12 @@
+import { InjectionToken } from '@angular/core';
+
+export interface Config {
+  extraSmall?: number;
+  small?: number;
+  medium?: number;
+  large?: number;
+  extraLarge?: number;
+  disableForTesting?: boolean;
+}
+
+export const NGX_EQ_CONFIG = new InjectionToken<Config>('NGX_EQ_CONFIG');
