+++ projects/ngx-eq/src/ngx-eq.module.ts
@@ -2,17 +2,9 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { NgxEqDirective } from './ngx-eq.directive';
+import { Config, NGX_EQ_CONFIG } from './ngx-eq.di-tokens';
 
-export interface Config {
-  extraSmall?: number;
-  small?: number;
-  medium?: number;
-  large?: number;
-  extraLarge?: number;
-  disableForTesting?: boolean;
-}
-
-export const defaultConfig = {
+export const defaultConfig: Config = {
   extraSmall: 280,
   small: 480,
   medium: 720,
@@ -27,12 +19,12 @@ export const defaultConfig = {
   exports: [NgxEqDirective]
 })
 export class NgxEqModule {
-  static forRoot(config = {}): ModuleWithProviders {
+  static forRoot(config: Config = {}): ModuleWithProviders {
     return {
       ngModule: NgxEqModule,
       providers: [
         {
-          provide: 'config',
+          provide: NGX_EQ_CONFIG,
           // inline config, otherwise they get treeshaken out when they shouldnt
           useValue: {
             extraSmall: 280,
